RIP 6 hours

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mechanisms/Sorting/QuickSpindexer.java
@@ -258,6 +258,6 @@ public class QuickSpindexer { // Prefix for commands
     }
 
     public static boolean aligned(){
-        return abs(spindexer.getCurrentPosition() - spindexer.getTargetPosition()) < errorMargin && spindexer.getTargetPosition() == targetPosition;
+        return abs(spindexer.getCurrentPosition() - spindexer.getTargetPosition()) < errorMargin && spindexer.getTargetPosition() == (int) targetPosition;
     }
 }
